moved most SubstrateAnnotationExtractor logic in com.oracle.svm.util
