+++ substratevm/src/com.oracle.graal.pointsto/src/com/oracle/graal/pointsto/meta/BaseLayerField.java
@@ -26,6 +26,8 @@ package com.oracle.graal.pointsto.meta;
 
 import java.lang.annotation.Annotation;
 
+import com.oracle.svm.util.AnnotationsContainer;
+
 import jdk.graal.compiler.debug.GraalError;
 import jdk.vm.ci.meta.JavaType;
 import jdk.vm.ci.meta.ResolvedJavaField;
@@ -37,7 +39,7 @@ import jdk.vm.ci.meta.ResolvedJavaType;
  * If a field cannot be looked up by name, a {@link BaseLayerField} is created and put in an
  * {@link AnalysisField} to represent this missing field, using the information from the base layer.
  */
-public class BaseLayerField extends BaseLayerElement implements ResolvedJavaField {
+public class BaseLayerField extends AnnotationsContainer implements ResolvedJavaField {
     private final int id;
     private final String name;
     private final ResolvedJavaType declaringClass;
