+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/cenum/CEnumCallWrapperMethod.java
@@ -82,7 +82,7 @@ public class CEnumCallWrapperMethod extends CustomSubstitutionMethod {
         if (AnnotationUtil.getAnnotation(original, CEnumValue.class) != null) {
             return INJECTED_ANNOTATIONS;
         }
-        return null;
+        return List.of();
     }
 
     @Override
